+++ sentry/src/main/java/io/sentry/protocol/User.java
@@ -34,7 +34,11 @@ public final class User implements JsonUnknown, JsonSerializable {
   /** Username of the user. */
   private @Nullable String username;
 
-  private @Nullable String segment;
+  /**
+   * @deprecated has no effect and will be removed in the next major update. Use a custom tag or
+   *     context instead.
+   */
+  @Deprecated private @Nullable String segment;
 
   /** Remote IP address of the user. */
   private @Nullable String ipAddress;
@@ -224,7 +228,9 @@ public final class User implements JsonUnknown, JsonSerializable {
    * Gets the segment of the user.
    *
    * @return the user segment.
+   * @deprecated has no effect and will be removed in the next major update.
    */
+  @Deprecated
   public @Nullable String getSegment() {
     return segment;
   }
@@ -233,7 +239,9 @@ public final class User implements JsonUnknown, JsonSerializable {
    * Sets the segment of the user.
    *
    * @param segment the segment.
+   * @deprecated has no effect and will be removed in the next major update.
    */
+  @Deprecated
   public void setSegment(final @Nullable String segment) {
     this.segment = segment;
   }